Fix dashboard block access checking #1031
Fixes the following error that occurs if access is denied to a dashboard
block:

InvalidArgumentException: "0" is an invalid render array key. Value should be
an array but got a string. in Drupal\Core\Render\Element::children()
(line 97 of core/lib/Drupal/Core/Render/Element.php).

Dashboard panes should always return a render array, but the default
$output value was a string. In the case of blocks that deny access, this
string was not being overwritten.